requirements.txt
alembic==1.11.0
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
cachelib==0.9.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Flask==2.3.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.19.3
pytest==7.3.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
qrcode==7.4.2
ratelim==0.1.6
requests==2.30.0
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
